chore(build): Run semantic release on latest node
When running the build on Node 6, we only have NPM 3, which doesn't have `npx`. So by adding the deploy `on` filter to run `semantic-release` on the latest node version, we won't try to even run it on Node 6.

BREAKING CHANGE: We need to trigger a version bump after taking over the package