chore(docs): Add use-case for empty string baseUrl
It was requested in #47 to add support to `.request()` for full URLs because [webooks](https://www.eventbrite.com/developer/v3/api_overview/webhooks/) returns full API urls. However instead of updating `.request()` a client can just set the `baseUrl` to empty string (`''`) when configuring the SDK object.

This adds docs and an example for future users.

Closes #47.